order controller 1st;

+++ src/main/java/com/ssmshop/service/impl/order/OrderServiceImpl.java
@@ -4,6 +4,7 @@ import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import com.ssmshop.bo.order.OrderListBo;
 import com.ssmshop.common.ServerResponse;
+import com.ssmshop.dao.GoodsMapper;
 import com.ssmshop.dao.OrderActionMapper;
 import com.ssmshop.dao.OrderGoodsMapper;
 import com.ssmshop.dao.OrderMapper;
@@ -17,6 +18,7 @@ import com.ssmshop.utils.ToVoUtils;
 import com.ssmshop.vo.order.OrderDetailVo;
 import com.ssmshop.vo.order.OrderListVo;
 import org.joda.time.DateTime;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -32,6 +34,8 @@ public class OrderServiceImpl extends BaseServiceImpl implements IOrderService {
     private OrderGoodsMapper orderGoodsMapper;
     @Autowired
     private OrderActionMapper orderActionMapper;
+    @Autowired
+    private GoodsMapper goodsMapper;
     @Override
     public ServerResponse<PageInfo<OrderListVo>> list(int pageNum, int pageSize, OrderListBo orderListBo) {
         OrderExample orderExample = OrderExampleExtends.getOrderExample(orderListBo);
@@ -45,16 +49,21 @@ public class OrderServiceImpl extends BaseServiceImpl implements IOrderService {
     }
 
     @Override
-    public ServerResponse<?> edit(int orderId, Order order,List<OrderGoods> orderGoodsList) {
+    public ServerResponse<?> edit(int orderId, Order order,String orderGoodsIds) {
         if (orderId >0){
             order.setOrderId(orderId);
-            OrderGoodsExample orderGoodsExample = new OrderGoodsExample();
-            orderGoodsExample.createCriteria().andOrderIdEqualTo(orderId);
-            this.orderGoodsMapper.deleteByExample(orderGoodsExample);
-            for (OrderGoods orderGoods: orderGoodsList
-                    ) {
-                orderGoods.setOrderId(orderId);
-                this.orderGoodsMapper.insertSelective(orderGoods);
+            if (!"".equals(orderGoodsIds)){
+                String[] orderGoodsIdArr = orderGoodsIds.split("_");
+                OrderGoodsExample orderGoodsExample = new OrderGoodsExample();
+                orderGoodsExample.createCriteria().andOrderIdEqualTo(orderId);
+                this.orderGoodsMapper.deleteByExample(orderGoodsExample);
+                for (int i= 0;i<orderGoodsIdArr.length;++i){
+                    Goods goods = this.goodsMapper.selectByPrimaryKey(Integer.parseInt(orderGoodsIdArr[i]));
+                    OrderGoods orderGoods = new OrderGoods();
+                    orderGoods.setOrderId(orderId);
+                    BeanUtils.copyProperties(goods,orderGoods);
+                    this.orderGoodsMapper.insertSelective(orderGoods);
+                }
             }
             return super.responseBase(this.orderMapper.updateByPrimaryKeySelective(order));
         }else{
@@ -64,14 +73,19 @@ public class OrderServiceImpl extends BaseServiceImpl implements IOrderService {
     }
 
     @Override
-    public ServerResponse<?> add(Order order,List<OrderGoods> orderGoodsList) {
+    public ServerResponse<?> add(Order order,String orderGoodsIds) {
         order.setOrderSn(DateTimeUtils.dateToNum(System.currentTimeMillis()));
-       int orderId = this.orderMapper.insertSelective(order);
-        for (OrderGoods orderGoods: orderGoodsList
-             ) {
+        this.orderMapper.insertSelective(order);
+       int orderId = order.getOrderId();
+        String[] orderGoodsIdArr = orderGoodsIds.split("_");
+        for (int i = 0;i<orderGoodsIdArr.length;++i){
+            OrderGoods orderGoods = new OrderGoods();
             orderGoods.setOrderId(orderId);
+            Goods goods = this.goodsMapper.selectByPrimaryKey(Integer.parseInt(orderGoodsIdArr[i]));
+            BeanUtils.copyProperties(goods,orderGoods);
             this.orderGoodsMapper.insertSelective(orderGoods);
         }
+
         if (orderId>0){
             return ServerResponse.createBySuccess();
         }else{
@@ -107,5 +121,11 @@ public class OrderServiceImpl extends BaseServiceImpl implements IOrderService {
             return ServerResponse.createByError();
         }
 
+    }
+
+    @Override
+    public ServerResponse<String> editOrderAction(OrderAction orderAction) {
+
+        return super.responseBase(this.orderActionMapper.updateByPrimaryKeySelective(orderAction));
     }
 }
